knowledge-base/requirements.txt
# knowledge-base/requirements.txt
# All packages approved in STD09 §2.4 — Decision record D01 (2026-03-14)
# Versions are pinned. Re-approval required before upgrading any version.

lancedb==0.29.2
sentence-transformers==5.3.0
python-frontmatter==1.1.0
pyyaml==6.0.3
